using BigDecimal string representations to store user spending

+++ app/src/main/java/edu/uga/cs/weliquid/EnterPriceDialogFragment.java
@@ -30,7 +30,7 @@ public class EnterPriceDialogFragment extends DialogFragment {
     // A class implementing this interface will handle the new shopping list item, i.e. store it
     // in Firebase and add it to the RecyclerAdapter.
     public interface EnterPriceDialogListener {
-        void enterItemPrice(float value);
+        void enterItemPrice(BigDecimal value);
     }
 
     @NonNull
@@ -71,7 +71,7 @@ public class EnterPriceDialogFragment extends DialogFragment {
         public void onClick(DialogInterface dialog, int which) {
             // get the price from the user
             String itemPrice = itemPriceView.getText().toString();
-            float actualPrice = Float.valueOf(itemPrice);
+            BigDecimal actualPrice = BigDecimal.valueOf(Double.valueOf(itemPrice));
 
             // get the Activity's listener to add the new shopping list item
             EnterPriceDialogListener listener = (EnterPriceDialogListener) getActivity();
